remove css modules and create identification component
