+++ src/components/IdentificationForms/index.css
@@ -1,0 +1,30 @@
+.form__wrapper {
+  display: flex;
+  width: min(100%, 500px);
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.form__wrapper > div {
+  margin-bottom: 1.5rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.form__wrapper > form {
+  background-color: white;
+  width: 100%;
+  padding: 1rem;
+  display: grid;
+  grid-template-rows: repeat(5, auto);
+  grid-row-gap: 1.5rem;
+}
+
+.form__footer {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
