--- src/containers/common/FluidGridContainer/types.d.ts
@@ -1,3 +1,0 @@
-export interface FluidContainerProps {
-  children: ReactChild | ReactChildren;
-}
