+++ src/containers/common/FormsContainer/types.d.ts
@@ -1,0 +1,8 @@
+import { IdentificationTypes } from "../../../components/IdentificationForms";
+
+export interface FormsContainerProps {
+  imageURL: string;
+  blurURL: string;
+  alt: string;
+  type: IdentificationTypes;
+};
