+++ src/containers/signIn/index.tsx
@@ -1,38 +1,9 @@
-import Grid from "@material-ui/core/Grid";
-
-import FluidGridContainer from "../common/FluidGridContainer";
-import SignInForm from "../../components/SignInForm";
-import AsideImage from "../../components/AsideImage";
-import Styles from "../../styles/common/index.module.css";
-
+import FormsContainer from "../common/FormsContainer";
+import { componentTypes } from "../../components/IdentificationForms";
 import { asideImageProps } from "./constants";
 
-const index = () => {
-  return (
-    <FluidGridContainer>
-      <Grid
-        item
-        lg={7}
-        md={4}
-        sm={false}
-        className={[Styles.overflow__hidden, Styles.hide__onSm].join(" ")}
-      >
-        <AsideImage {...asideImageProps} />
-      </Grid>
-      <Grid
-        item
-        lg={5}
-        md={8}
-        sm={12}
-        container
-        direction="row"
-        justifyContent="center"
-        alignItems="center"
-      >
-        <SignInForm />
-      </Grid>
-    </FluidGridContainer>
-  );
-};
+const index = () => (
+  <FormsContainer {...asideImageProps} type={componentTypes.signIn.id} />
+);
 
 export default index;
